rename auth.ts to authRoutes.ts for consistency
