+++ backend/src/routes/authRoutes.ts
